requirements.txt
colorlog==6.9.0
homeassistant==2025.2.4
pip>=21.3.1
ruff==0.11.2
xmltodict==0.14.2

# Testing
pytest-asyncio
